requirements.txt
sanic==21.9.1
sanic-cors==1.0.1
sanic-openapi==21.6.1
termcolor==1.1.0
jsonschema==3.2.0
python-dotenv==0.19.2
urllib3==1.26.7
cryptoaddress==0.2.1
pymongo==3.11.4
click==8.0.3
sanic-ext==21.12.1
PyJWT==2.3.0
requests==2.27.1
websockets==9.1
web3==5.28.0
python-telegram-bot==13.7
python-rocksdb==0.7.0